adding topics and styling

+++ client/app/admin/admin.component.ts
@@ -48,11 +48,26 @@ export class AdminComponent implements OnInit {
   }
 
   deleteUser(user) {
-    this.userService.deleteUser(user).subscribe(
-      data => this.toast.setMessage('user deleted successfully.', 'success'),
-      error => console.log(error),
-      () => this.getUsers()
-    );
+    if (window.confirm('Are you sure you want to permanently delete this user?')) {
+      this.userService.deleteUser(user).subscribe(
+        data => this.toast.setMessage('user deleted successfully.', 'success'),
+        error => console.log(error),
+        () => this.getUsers()
+      );
+    } 
+  }
+
+  deleteTopic(topic) {
+    if (window.confirm('Are you sure you want to permanently delete this item?')) {
+      this.topicService.deleteTopic(topic).subscribe(
+        res => {
+          const pos = this.topics.map(elem => elem._id).indexOf(topic._id);
+          this.topics.splice(pos, 1);
+          this.toast.setMessage('item deleted successfully.', 'success');
+        },
+        error => console.log(error)
+      );
+    }
   }
 
   getTopics() {
@@ -85,6 +100,10 @@ export class AdminComponent implements OnInit {
       error => console.log(error),
       () => console.log('categories loaded')
     );
+  }
+
+  save_reordering(){
+    console.log(this.topics);
   }
 
     // Dialog for editing topics
